rviz: simplified API of RenderPanel by removing need for RenderSystem::get() and showEvent() override in subclass.  This will break plugins which use RenderPanels, they will need to be updated to match the new API.
git-svn-id: https://code.ros.org/svn/ros-pkg/stacks/visualization/trunk@38640 eb33c2ac-9c88-4c90-87e0-44a10359b0c3

+++ src/test/mesh_marker_test.cpp
@@ -14,7 +14,7 @@ void publishMesh( int id, float x, float y, float r, float g, float b, float a, 
   marker.header.frame_id = "/base_link";
   marker.header.stamp = ros::Time::now();
   marker.ns = "mesh";
-  marker.type = mesh ? visualization_msgs::Marker::MESH_RESOURCE : visualization_msgs::Marker::SPHERE;
+  marker.type = mesh ? (int) visualization_msgs::Marker::MESH_RESOURCE : (int) visualization_msgs::Marker::SPHERE;
   marker.action = visualization_msgs::Marker::ADD;
   marker.pose.orientation.x = 0.0;
   marker.pose.orientation.y = 0.0;
